perf: Handroll Deserialize implementations for CompletionResponse and CompletionTextEdit
In Zed, we've noticed that deserialization of CompletionResponse takes non-trivial amount of time (it takes ~380ms to deserialize a 60MB of JSON). We've attributed it to #[serde(untagged)] and rewrote the Deserialize impls manually, cutting down the time for the same scenario down to 120ms. The impls are actually pretty straightforward, as the types being deserialized have a distinct structure to them which allows us to dispatch to other Deserialize impl on the Visitor level.